Track a tree of where Hypothesis has explored
This lets us do two important things:

* We can avoid testing duplicate buffers, both during generation
  and shrinking.
* When shrinking we can additionally avoid testing any buffers
  that are a prefix of a buffer we've already tried (which will
  result in an overrun).

This is partly intended as a performance improvement, partly intended
as a data quality improvement. The data quality improvement is
straightforward: In cases where we previously wasted our example
budget on duplicates, we now don't.

The performance improvement in our
own tests is mild to non-existent, but that's because the actual
functions under test are quite fast. For things where that's not
the case, this should be a significant improvement. I especially
expect it to be useful for stateful testing.

Another small change this makes is that we no longer seed the
global random module to an interesting seed by default. This is
because that creates a way for values to look distinct where
they are not. Instead we just seed it to zero, and rely on the
user to ask for it if they want a more interesting seed.